Using sorted sets

+++ domain/base.go
@@ -2,7 +2,7 @@ package domain
 
 type BaseRepository interface {
 	NewId(fields ...string) string
-	CountAll() (int, error)
+	CountAll() (int64, error)
 	Exists(id string) (bool, error)
 }
